kkp-pinterest-scheduler/src/requirements.txt
python-pinterest-api==5.0.0
pandas==2.1.0
python-dotenv==1.0.0
paramiko==3.3.1
SQLAlchemy==2.0.21
schedule==1.2.0
python-crontab==3.0.0